CookiesDb: 1. 異動 Products.Price 欄位資料型態, 2. 加入 View

+++ CookiesDb/dbo/Tables/Products.sql
@@ -2,7 +2,7 @@
     [ProductId]   INT           NOT NULL,
     [ProductName] NVARCHAR (30) NOT NULL,
     [CategoryId]  INT           NOT NULL,
-    [Price]       INT           NOT NULL,
+    [Price]       DECIMAL(6, 2)           NOT NULL,
     CONSTRAINT [PK_Products] PRIMARY KEY CLUSTERED ([ProductId] ASC),
     CONSTRAINT [FK_Categories] FOREIGN KEY ([CategoryId]) REFERENCES [dbo].[Categories] ([CategoryId])
 );
